[HUMAN App] refactor: reset password services (#3180)
