+++ packages/apps/human-app/frontend/src/modules/worker/reset-password/hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './reset-password';
