+++ packages/apps/human-app/frontend/src/modules/worker/reset-password/reset-password-success.page.tsx
