+++ packages/apps/human-app/frontend/src/modules/worker/send-reset-link/index.ts
@@ -1,0 +1,2 @@
+export * from './send-reset-link-success.page';
+export * from './send-reset-link.page';
